Replace select2 styling
Previously, we styled select2 elements to be very blue (under default
colours). This didn't match the visual language we used for the rest of
our inputs, which have a plain white background. The select2 inputs had
a style more similar to buttons, which hold a different purpose.

This commit replaces the existing select2 styling with styling meant to
match the look of a standard select element with bootstrap's styling.

This looks very similar to select2 v4's styling, and we had previously
intended to switch to select2 v4's default styling after upgrading, but
it no longer looks like we will be upgrading.