Do not set order state before redirecting
During checkout we ensure a valid state be checking the state in the params
is actually a valid checkout step and if not redirect to first step of checkout.

Though, right before redirecting we set the state attribute of the order to cart.
This does not have any effect at all, as the redirect causes the controller to reload
the order on the next request anyway.