Introduce PaymentProvider class
Currently, the Solidus codebase conflates the idea of
a Gateway and a Payment Method. This adds a payment provider
class to store credentials and to be able to query payments
by payment provider account.

It also projects moving the credentials for payment methods
from the payment method to the provider.

This way, we can succinctly model payment providers with
several payment methods, such as Braintree, Stripe or Adyen,
without having to duplicate credentials.