Always show "Continue" on admin new payment screen
Former we displayed "Continue" when the order is in the cart state and "Update" when not. Since this is the Payments create screen we should never display "Update". Ideally we would display "Create", but since this screen is often part of an admin checkout flow "Continue" is best bet here.