Just touch taxons in Product#touch_taxons
And make use of the `after_touch` callback of `Spree::Taxon`,
that already handles touching ancestors and its taxonomy.

This hopefully fixes #3931 as well, because it runs each `touch`
callback in it's own transaction instead of one large transaction
that might cause dead locks.

This IS slower than the optimized approach before, but this should not
be that much of an issue, since we are updating a product here
and this will a) not happen that much and b) there will not be many taxons
one single product will be attached to, right? Even in large stores this
will most likely be in single to double digits, instead of hundreds
or thousands. But I might be painfully wrong here.

This also has the advantage that a `after_touch` callback a store or an
extenstion might have will be triggered and caches will be invalidated
correctly.